requirements.txt
streamlit
pandas
langchain>=0.1.17
langchain-community>=0.0.20
openai
python-dotenv
